+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Changed
 
+- Moved the `generateStaticFiles` method outside of `index.ts`.
 - Moved the `clean` method in `index.ts` to its own utility.
 - Upgraded Node.js to `20.10.0` and npm to `10.2.3`.
 - Set up a monorepo structure.